Refactor how IdP errors are parsed from WebView
