+++ Pod/Classes/Core/Private/NSError+A0APIError.h
@@ -21,6 +21,7 @@
 // THE SOFTWARE.
 
 #import <Foundation/Foundation.h>
+#import "A0ErrorCode.h"
 
 @interface NSError (A0APIError)
 
@@ -33,5 +34,15 @@
 - (NSString *)a0_code;
 
 - (NSError *)a0_errorWithPayload:(NSDictionary *)payload;
+
+- (BOOL)a0_auth0ErrorWithCode:(A0ErrorCode)code;
+
+- (BOOL)a0_cancelledSocialAuthenticationError;
+
++ (NSError *)errorWithCode:(NSInteger)code description:(NSString *)description payload:(NSDictionary *)payload;
+
++ (NSError *)errorWithCode:(NSInteger)code userInfo:(NSDictionary *)userInfo;
+
++ (NSError *)errorWithCode:(NSInteger)code description:(NSString *)description failureReason:(NSString *)failureReason;
 
 @end
