+++ Pod/Classes/Core/Private/NSError+A0APIError.m
@@ -23,6 +23,7 @@
 #import "NSError+A0APIError.h"
 
 NSString * const A0JSONResponseSerializerErrorDataKey = @"A0JSONResponseSerializerErrorDataKey";
+NSString * const A0ErrorDomain = @"com.auth0";
 
 @implementation NSError (A0APIError)
 
@@ -46,6 +47,40 @@ NSString * const A0JSONResponseSerializerErrorDataKey = @"A0JSONResponseSerializ
     NSMutableDictionary *userInfo = [self.userInfo mutableCopy];
     userInfo[A0JSONResponseSerializerErrorDataKey] = payload;
     return [NSError errorWithDomain:self.domain code:self.code userInfo:[NSDictionary dictionaryWithDictionary:userInfo]];
+}
+
+- (BOOL)a0_auth0ErrorWithCode:(A0ErrorCode)code {
+    return [self.domain isEqualToString:A0ErrorDomain] && self.code == code;
+}
+
+- (BOOL)a0_cancelledSocialAuthenticationError {
+    return self.code == A0ErrorCodeFacebookCancelled
+    || self.code == A0ErrorCodeTwitterCancelled
+    || self.code == A0ErrorCodeAuth0Cancelled
+    || self.code == A0ErrorCodeGooglePlusCancelled;
+}
+
++ (NSError *)errorWithCode:(NSInteger)code description:(NSString *)description payload:(NSDictionary *)payload {
+    return [self errorWithCode:code
+                      userInfo:@{
+                                 NSLocalizedDescriptionKey: description,
+                                 A0JSONResponseSerializerErrorDataKey: payload ?: @{},
+                                 }];
+}
+
++ (NSError *)errorWithCode:(NSInteger)code userInfo:(NSDictionary *)userInfo {
+    NSError *error = [NSError errorWithDomain:A0ErrorDomain
+                                         code:code
+                                     userInfo:userInfo];
+    return error;
+}
+
++ (NSError *)errorWithCode:(NSInteger)code description:(NSString *)description failureReason:(NSString *)failureReason {
+    return [self errorWithCode:code
+                      userInfo:@{
+                                 NSLocalizedDescriptionKey: description,
+                                 NSLocalizedFailureReasonErrorKey: failureReason,
+                                 }];
 }
 
 @end
