+++ Pod/Classes/TouchID/Private/A0TouchIDSignUpViewController.m
@@ -33,6 +33,7 @@
 #import "A0EmailValidator.h"
 #import "A0Lock.h"
 #import "NSObject+A0APIClientProvider.h"
+#import "NSError+A0APIError.h"
 
 @interface A0TouchIDSignUpViewController ()
 
@@ -94,8 +95,8 @@ AUTH0_DYNAMIC_LOGGER_METHODS
                            } failure:^(NSError *error){
                                @strongify(self);
                                [self.signUpButton setInProgress:NO];
-                               NSString *title = [A0Errors isAuth0Error:error withCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedDescription : A0LocalizedString(@"There was an error signing up");
-                               NSString *message = [A0Errors isAuth0Error:error withCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedFailureReason : [A0Errors localizedStringForSignUpError:error];
+                               NSString *title = [error a0_auth0ErrorWithCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedDescription : A0LocalizedString(@"There was an error signing up");
+                               NSString *message = [error a0_auth0ErrorWithCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedFailureReason : [A0Errors localizedStringForSignUpError:error];
                                A0ShowAlertErrorView(title, message);
                            }];
     } else {
