+++ Pod/Classes/UI/Private/A0ActiveDirectoryViewController.m
@@ -45,6 +45,7 @@
 #import "A0Lock.h"
 #import "NSObject+A0APIClientProvider.h"
 #import "NSObject+A0AuthenticatorProvider.h"
+#import "NSError+A0APIError.h"
 
 @interface A0ActiveDirectoryViewController ()
 
@@ -125,8 +126,8 @@ AUTH0_DYNAMIC_LOGGER_METHODS
                     @strongify(self);
                     [self postLoginErrorNotificationWithError:error];
                     [self.accessButton setInProgress:NO];
-                    NSString *title = [A0Errors isAuth0Error:error withCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedDescription : A0LocalizedString(@"There was an error logging in");
-                    NSString *message = [A0Errors isAuth0Error:error withCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedFailureReason : [A0Errors localizedStringForLoginError:error];
+                    NSString *title = [error a0_auth0ErrorWithCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedDescription : A0LocalizedString(@"There was an error logging in");
+                    NSString *message = [error a0_auth0ErrorWithCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedFailureReason : [A0Errors localizedStringForLoginError:error];
                     A0ShowAlertErrorView(title, message);
                 };
                 A0AuthParameters *parameters = self.parameters.copy;
@@ -172,6 +173,9 @@ AUTH0_DYNAMIC_LOGGER_METHODS
 - (void)loginUserWithConnection:(A0Connection *)connection {
     @weakify(self);
     [self.accessButton setInProgress:YES];
+
+    NSString *connectionName = connection.name;
+
     A0APIClientAuthenticationSuccess successBlock = ^(A0UserProfile *profile, A0Token *token){
         @strongify(self);
         [self postLoginSuccessfulForConnection:connection];
@@ -185,7 +189,7 @@ AUTH0_DYNAMIC_LOGGER_METHODS
         @strongify(self);
         [self postLoginErrorNotificationWithError:error];
         [self.accessButton setInProgress:NO];
-        if (![A0Errors isCancelledSocialAuthentication:error]) {
+        if (![error a0_cancelledSocialAuthenticationError]) {
             switch (error.code) {
                 case A0ErrorCodeTwitterAppNotAuthorized:
                 case A0ErrorCodeTwitterInvalidAccount:
@@ -198,7 +202,7 @@ AUTH0_DYNAMIC_LOGGER_METHODS
                     A0ShowAlertErrorView(error.localizedDescription, error.localizedFailureReason);
                     break;
                 default:
-                    A0ShowAlertErrorView(A0LocalizedString(@"There was an error logging in"), [A0Errors localizedStringForSocialLoginError:error]);
+                    A0ShowAlertErrorView(A0LocalizedString(@"There was an error logging in"), [A0Errors localizedStringForConnectionName:connectionName loginError:error]);
                     break;
             }
         }
@@ -206,7 +210,6 @@ AUTH0_DYNAMIC_LOGGER_METHODS
 
     A0IdentityProviderAuthenticator *authenticator = [self a0_identityAuthenticatorFromProvider:self.lock];
     A0AuthParameters *parameters = self.parameters.copy;
-    NSString *connectionName = connection.name;
     A0LogVerbose(@"Authenticating with connection %@", connectionName);
     [authenticator authenticateWithConnectionName:connectionName parameters:parameters success:successBlock failure:failureBlock];
 }
