+++ Pod/Classes/UI/Private/A0ChangePasswordViewController.m
@@ -44,6 +44,7 @@
 #import "A0ConfirmPasswordValidator.h"
 #import "A0Lock.h"
 #import "NSObject+A0APIClientProvider.h"
+#import "NSError+A0APIError.h"
 
 @interface A0ChangePasswordViewController ()
 
@@ -152,8 +153,8 @@
             @strongify(self);
             [self postChangePasswordErrorNotificationWithError:error];
             [self.recoverButton setInProgress:NO];
-            NSString *title = [A0Errors isAuth0Error:error withCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedDescription : A0LocalizedString(@"Couldn't change your password");
-            NSString *message = [A0Errors isAuth0Error:error withCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedFailureReason : [A0Errors localizedStringForChangePasswordError:error];
+            NSString *title = [error a0_auth0ErrorWithCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedDescription : A0LocalizedString(@"Couldn't change your password");
+            NSString *message = [error a0_auth0ErrorWithCode:A0ErrorCodeNotConnectedToInternet] ? error.localizedFailureReason : [A0Errors localizedStringForChangePasswordError:error];
             A0ShowAlertErrorView(title, message);
         };
         A0APIClient *client = [self a0_apiClientFromProvider:self.lock];
